backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
pydantic==2.6.4
easyocr==1.7.1
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
rapidfuzz==3.6.2
Pillow==10.2.0
python-dotenv==1.0.1
deep-translator==1.11.4
